Update image stamp with external timestamp
Subscribe to /extern_timestamp topic to get shutter timestamp from an
external source. The callback sets a ros::Time variable, extern_stamp.

If extern_stamp is non-zero, the difference between the external
timestamp and the image header timestamp for the NEXT image received
is computed. extern_stamp is then reset to zero but the calculated
difference remains. This is repeated every time a new external timestamp
is received. However, only one external timestamp is required to
calculate the offset and provide accurate timestamps.

If the difference variable is non-zero, the difference is applied to
the header timestamp of all incoming images. This becomes the new
timestamp.

If an external timestamp is not received, the difference variable will
remain at zero and the header timestamp for images will instead be
set to ros::Time.now() (for backwards compatibility).