Add recv_ready topic for time synchronization
At bootup, once connection with the camera has been established, send
an empty message to the recv_ready topic. The time sychronization
module can listen to this topic and interpret this to mean that a time
sync is required.
This is done at bootup only. Consecutive connects/disconnects do not
publish to this thread. The assumption being that the camera will not be
powered OFF.